revert error codes and remove previous implementations
